TEAM C/Transformer/김종윤/김종윤/requirements.txt
spacy==2.1.8
soynlp
torchtext==0.4.0